Remove DLite as a dependency
Validate via the pydantic SOFT5 and SOFT7 models instead of trying to
instantiate entities in DLite.

This greatly simplifies things and automatically implements Python
3.12 support for the CLI.

Add helpful READMEs in the static folders for valid and invalid entities
to describe them.
This was done especially to showcase the validation errors pydantic
throws for the invalid entity Person.json and to announce the Cat.json
valid entity as a SOFT5 entity (the others are SOFT7 entities).