Implement new logic to check possible collisions
The new logic for the upload CLI command is:

Check if the incoming entity's URI already exists in the backend.
If no, upload.
Otherwise, check if the incoming entity is equal in every way to the
existing entity in the backend.
If yes, skip and continue.
Otherwise, have the user confirm if they want to upload the incoming
entity with a new version.
If no, skip and continue.
If yes, ask the user to enter the new version, while providing a
reasonable default value.
If user aborts, skip and continue.
Otherwise, update the incoming entity with the new version and upload
it.

We also add a new entity with a "ref" property type, since we need to
make sure special keys can be uploaded to the backend (Mongo does not
support keys starting with $).