Add authSource and authMethod to client if using x509
Add tests for the /_admin/create endpoint.

Avoid running tests with a live backend for most of the CLI tests.
This is due to the fact that authentication is now done via OAuth2
against an external service. A mock for this has not been put in place.
It could, in theory, be done, since the target OAuth2 base URL/hostname
is a configurable str, but this requires either another Docker service
to run just for OAuth2 - or a well-written socket-mocked server to do
the OAuth2 handshaking.