Fully implement list CLI cmd
list namespaces

Will list all the namespaces as a full namespace, valid to be a value in
an entity under the namespace key.

list entities

Will list all entities for a given namespace or multiple namespaces or
all namespaces depending on the provided arguments (if any) or if the
`-a/--all` option is provided.

This has all been made possible through the (normally hidden) endpoints
`/_api/namespaces` and `/_api/entities`. Both are GET endpoints.

The endpoints will list all namespaces or all entities with a given list
of namespaces (provided through the query parameter 'namespace', may be
supplied multiple times - or no times).